Allow marking methods as NOT involved in test discovery
i.e. not an examples method, not an assertion method, etc. etc.

Example (with_context_manager is just a regular method and is ignored by
contexts except to record that fact -- in particular it is not treated as a
setup method despite the word "context" appearing in the method name):

    from contexts import method

    class When_whatever:

        @method
        def with_context_manager(self):
            ...

        def should_whatever(self):
            ...